requirements.txt
asgiref==3.7.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
defusedxml==0.7.1
Django==5.0.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
dotenv==0.9.9
idna==3.10
oauthlib==3.2.2
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
python3-openid==3.2.0
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.4.4
stripe==11.6.0
typing_extensions==4.10.0
urllib3==2.3.0
